requirements-dev.txt
autopep8==1.7.0
black==22.8.0
isort==5.10.1


Werkzeug[watchdog]==3.0.3
ipdb==0.13.13
psycopg[binary]==3.2.1

mypy==1.11.1
django-stubs[compatible-mypy]==5.0.4
pytest==8.3.2
pytest-sugar==1.0.0
djangorestframework-stubs==3.15.0


sphinx==7.4.7
sphinx-autobuild==2024.4.16

coverage==7.6.0
pre-commit==3.8.0


factory-boy==3.3.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-coverage-plugin==3.1.0
pytest-django==4.8.0
